+++ feature/main-nav-graph/src/main/kotlin/org/bmsk/lifemash/feature/nav/LifeMashNavController.kt
@@ -3,6 +3,6 @@ package org.bmsk.lifemash.feature.nav
 import androidx.navigation.NavController
 
 interface LifeMashNavController<T> {
-    fun route(): String = ""
+    fun route(): String
     fun navigate(navController: NavController, navInfo: T)
 }
