+++ feature/main/build.gradle.kts
@@ -11,11 +11,12 @@ android {
 }
 
 dependencies {
-    implementation(projects.feature.topic)
     implementation(projects.feature.all)
 
     implementation(projects.feature.mainNavGraph)
     implementation(projects.feature.scrapApi)
+    implementation(projects.feature.topicApi)
+    implementation(projects.feature.webviewApi)
 
     // navigation
     implementation(libs.androidx.navigation.ui.ktx)
