+++ feature/main/src/main/kotlin/org/bmsk/lifemash/feature/main/MainActivity.kt
@@ -1,6 +1,7 @@
 package org.bmsk.lifemash.feature.main
 
 import android.os.Bundle
+import android.webkit.WebView
 import android.widget.Toast
 import androidx.activity.OnBackPressedCallback
 import androidx.activity.compose.setContent
@@ -8,8 +9,13 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.core.app.ActivityCompat
 import dagger.hilt.android.AndroidEntryPoint
 import org.bmsk.lifemash.core.designsystem.theme.LifeMashTheme
+import org.bmsk.lifemash.feature.nav.LifeMashNavController
 import org.bmsk.lifemash.feature.scrap.api.ScrapNavController
 import org.bmsk.lifemash.feature.scrap.api.ScrapNavGraph
+import org.bmsk.lifemash.feature.topic.api.TopicNavController
+import org.bmsk.lifemash.feature.topic.api.TopicNavGraph
+import org.bmsk.lifemash.feature.topic.api.WebViewNavController
+import org.bmsk.lifemash.feature.topic.api.WebViewNavGraph
 import javax.inject.Inject
 import kotlin.system.exitProcess
 
@@ -22,8 +28,21 @@ internal class MainActivity : AppCompatActivity() {
 
     @Inject
     lateinit var scrapNavGraph: ScrapNavGraph
+
     @Inject
     lateinit var scrapNavController: ScrapNavController
+
+    @Inject
+    lateinit var topicNavGraph: TopicNavGraph
+
+    @Inject
+    lateinit var topicNavController: TopicNavController
+
+    @Inject
+    lateinit var webViewNavGraph: WebViewNavGraph
+
+    @Inject
+    lateinit var webViewNavController: WebViewNavController
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -32,12 +51,16 @@ internal class MainActivity : AppCompatActivity() {
         setContent {
             LifeMashTheme {
                 val navigator = rememberMainNavigator(
-                    scrapNavController = scrapNavController
+                    scrapNavController = scrapNavController,
+                    topicNavController = topicNavController,
+                    webViewNavController = webViewNavController,
                 )
 
                 MainScreen(
                     navigator = navigator,
                     scrapNavGraph = scrapNavGraph,
+                    topicNavGraph = topicNavGraph,
+                    webViewNavGraph = webViewNavGraph,
                 )
             }
         }
