+++ feature/main/src/main/kotlin/org/bmsk/lifemash/feature/main/MainNavigator.kt
@@ -7,34 +7,46 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
 import org.bmsk.lifemash.feature.scrap.api.ScrapNavController
-import org.bmsk.lifemash.feature.topic.navigation.TopicRoute
-import org.bmsk.lifemash.feature.topic.navigation.navigateWebView
+import org.bmsk.lifemash.feature.topic.api.TopicNavController
+import org.bmsk.lifemash.feature.topic.api.WebViewNavController
+import org.bmsk.lifemash.feature.topic.api.WebViewNavControllerInfo
 
 internal class MainNavigator(
-    val navController: NavHostController,
     private val scrapNavController: ScrapNavController,
+    private val topicNavController: TopicNavController,
+    private val webViewNavController: WebViewNavController,
+    val navController: NavHostController,
 ) {
     private val currentDestination: NavDestination?
-        @Composable get() = navController.currentBackStackEntryAsState().value?.destination
+        @Composable get() = navController
+            .currentBackStackEntryAsState().value?.destination
 
-    val startDestination = TopicRoute.route
+    val startDestination = topicNavController.route()
 
     fun navigateWebView(url: String) {
-        navController.navigateWebView(url)
+        webViewNavController.navigate(navController, WebViewNavControllerInfo(url))
     }
 
     fun navigateScrap() {
         scrapNavController.navigate(navController, Unit)
+    }
+
+    fun navigateTopic() {
+        topicNavController.navigate(navController, Unit)
     }
 }
 
 @Composable
 internal fun rememberMainNavigator(
     scrapNavController: ScrapNavController,
+    topicNavController: TopicNavController,
+    webViewNavController: WebViewNavController,
     navController: NavHostController = rememberNavController(),
 ): MainNavigator = remember(navController) {
     MainNavigator(
         scrapNavController = scrapNavController,
+        topicNavController = topicNavController,
+        webViewNavController = webViewNavController,
         navController = navController
     )
 }
