+++ feature/scrap/src/main/kotlin/org/bmsk/lifemash/feature/scrap/ScrapScreen.kt
@@ -1,12 +1,14 @@
 package org.bmsk.lifemash.feature.scrap
 
 import android.content.res.Configuration
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
 import kotlinx.collections.immutable.PersistentList
 import kotlinx.collections.immutable.persistentListOf
 import org.bmsk.lifemash.core.designsystem.theme.LifeMashTheme
@@ -18,7 +20,10 @@ import org.bmsk.lifemash.feature.scrap.component.rememberScrapNewsItemState
 @Composable
 internal fun ScrapScreen(scrapNewsList: PersistentList<NewsModel>) {
     val scrapNewsItemState = rememberScrapNewsItemState()
-    LazyColumn(modifier = Modifier.fillMaxSize()) {
+    LazyColumn(
+        modifier = Modifier.fillMaxSize(),
+        verticalArrangement = Arrangement.spacedBy(12.dp),
+    ) {
         items(scrapNewsList) { news ->
             ScrapNewsItem(
                 newsModel = news,
@@ -41,43 +46,43 @@ private fun ScrapScreenPreview() {
                     NewsModel(
                         title = "NewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNews",
                         link = "",
-                        pubDate = DateParser.parseDate("2024-06-06"),
+                        pubDate = DateParser.parseDate("Tue, 20 Jun 2023 02:57:43 GMT"),
                         imageUrl = "",
                     ),
                     NewsModel(
                         title = "NewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNews",
                         link = "",
-                        pubDate = DateParser.parseDate("2024-06-06"),
+                        pubDate = DateParser.parseDate("Tue, 20 Jun 2023 02:57:43 GMT"),
                         imageUrl = "",
                     ),
                     NewsModel(
                         title = "NewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNews",
                         link = "",
-                        pubDate = DateParser.parseDate("2024-06-06"),
+                        pubDate = DateParser.parseDate("Tue, 20 Jun 2023 02:57:43 GMT"),
                         imageUrl = "",
                     ),
                     NewsModel(
                         title = "NewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNews",
                         link = "",
-                        pubDate = DateParser.parseDate("2024-06-06"),
+                        pubDate = DateParser.parseDate("Tue, 20 Jun 2023 02:57:43 GMT"),
                         imageUrl = "",
                     ),
                     NewsModel(
                         title = "NewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNews",
                         link = "",
-                        pubDate = DateParser.parseDate("2024-06-06"),
+                        pubDate = DateParser.parseDate("Tue, 20 Jun 2023 02:57:43 GMT"),
                         imageUrl = "",
                     ),
                     NewsModel(
                         title = "NewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNews",
                         link = "",
-                        pubDate = DateParser.parseDate("2024-06-06"),
+                        pubDate = DateParser.parseDate("Tue, 20 Jun 2023 02:57:43 GMT"),
                         imageUrl = "",
                     ),
                     NewsModel(
                         title = "NewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNewsNews",
                         link = "",
-                        pubDate = DateParser.parseDate("2024-06-06"),
+                        pubDate = DateParser.parseDate("Tue, 20 Jun 2023 02:57:43 GMT"),
                         imageUrl = "",
                     ),
                 ),
