+++ feature/topic-api/src/main/kotlin/org/bmsk/lifemash/feature/topic/api/TopicNavController.kt
@@ -1,0 +1,5 @@
+package org.bmsk.lifemash.feature.topic.api
+
+import org.bmsk.lifemash.feature.nav.LifeMashNavController
+
+interface TopicNavController : LifeMashNavController<Unit>
