+++ feature/topic/build.gradle.kts
@@ -7,6 +7,9 @@ android {
 }
 
 dependencies {
+    implementation(projects.feature.topicApi)
+    implementation(projects.feature.mainNavGraph)
+
     implementation(libs.androidx.appcompat)
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.activity)
