+++ feature/webview-api/src/main/kotlin/org/bmsk/lifemash/feature/topic/api/WebViewNavController.kt
@@ -1,0 +1,9 @@
+package org.bmsk.lifemash.feature.topic.api
+
+import org.bmsk.lifemash.feature.nav.LifeMashNavController
+
+interface WebViewNavController : LifeMashNavController<WebViewNavControllerInfo>
+
+data class WebViewNavControllerInfo(
+    val url: String,
+)
