+++ feature/webview-api/src/main/kotlin/org/bmsk/lifemash/feature/topic/api/WebViewNavGraph.kt
@@ -1,0 +1,9 @@
+package org.bmsk.lifemash.feature.topic.api
+
+import org.bmsk.lifemash.feature.nav.LifeMashNavGraph
+
+interface WebViewNavGraph : LifeMashNavGraph<WebViewNavGraphInfo>
+
+data class WebViewNavGraphInfo(
+    val onShowErrorSnackbar: (Throwable?) -> Unit,
+)
