+++ feature/webview/src/main/kotlin/org/bmsk/feature/webview/WebViewRoute.kt
@@ -1,0 +1,17 @@
+package org.bmsk.feature.webview
+
+import androidx.compose.runtime.Composable
+
+internal object WebViewRoute {
+    const val ROUTE = "webview"
+
+    fun createWebViewRoute(url: String): String = "$ROUTE?url=$url"
+
+    @Composable
+    operator fun invoke(
+        url: String,
+    ) {
+        WebViewScreen(url = url)
+    }
+}
+
